+++ lib/src/native/realm_core.dart
@@ -23,6 +23,7 @@ import 'dart:typed_data';
 
 // Hide StringUtf8Pointer.toNativeUtf8 and StringUtf16Pointer since these allows silently allocating memory. Use toUtf8Ptr instead
 import 'package:ffi/ffi.dart' hide StringUtf8Pointer, StringUtf16Pointer;
+import 'package:logging/logging.dart';
 
 import '../app.dart';
 import '../collections.dart';
@@ -1044,13 +1045,36 @@ class _RealmCore {
     });
   }
 
+  static void _logCallback(Pointer<Void> userdata, int levelAsInt, Pointer<Int8> message) {
+    try {
+      final logger = Realm.logger;
+      final level = _LogLevel.values[levelAsInt].loggerLevel;
+
+      // Don't do expensive utf8 to utf16 conversion unless we have to..
+      if (logger.isLoggable(level)) {
+        logger.log(level, message.cast<Utf8>().toDartString());
+      }
+    } finally {
+      _realmLib.realm_free(message.cast()); // .. but always free the message
+    }
+  }
+
   SyncClientConfigHandle _createSyncClientConfig(AppConfiguration configuration) {
     return using((arena) {
       final handle = SyncClientConfigHandle._(_realmLib.realm_sync_client_config_new());
 
       _realmLib.realm_sync_client_config_set_base_file_path(handle._pointer, configuration.baseFilePath.path.toUtf8Ptr(arena));
       _realmLib.realm_sync_client_config_set_metadata_mode(handle._pointer, configuration.metadataPersistenceMode.index);
-      _realmLib.realm_sync_client_config_set_log_level(handle._pointer, configuration.logLevel.index);
+      
+      _realmLib.realm_sync_client_config_set_log_level(handle._pointer, _LogLevel.fromLevel(Realm.logger.level).index);
+      _realmLib.realm_dart_sync_client_config_set_log_callback(
+        handle._pointer,
+        Pointer.fromFunction(_logCallback),
+        nullptr,
+        nullptr,
+        scheduler.handle._pointer,
+      );
+      
       _realmLib.realm_sync_client_config_set_connect_timeout(handle._pointer, configuration.maxConnectionTimeout.inMicroseconds);
       if (configuration.metadataEncryptionKey != null && configuration.metadataPersistenceMode == MetadataPersistenceMode.encrypted) {
         _realmLib.realm_sync_client_config_set_metadata_encryption_key(handle._pointer, configuration.metadataEncryptionKey!.toUint8Ptr(arena));
@@ -2001,5 +2025,28 @@ extension on realm_object_id {
       buffer[i] = bytes[i];
     }
     return ObjectId.fromBytes(buffer);
+  }
+}
+
+// Helper enum for converting Level
+enum _LogLevel {
+  all(RealmLogLevel.all),
+  trace(RealmLogLevel.trace),
+  debug(RealmLogLevel.debug),
+  detail(RealmLogLevel.detail),
+  info(RealmLogLevel.info),
+  warn(RealmLogLevel.warn),
+  error(RealmLogLevel.error),
+  fatal(RealmLogLevel.fatal),
+  off(RealmLogLevel.off);
+
+  final Level loggerLevel;
+  const _LogLevel(this.loggerLevel);
+
+  factory _LogLevel.fromLevel(Level level) {
+    for (final candidate in _LogLevel.values) {
+      if (level.value > candidate.loggerLevel.value) return candidate;
+    }
+    return _LogLevel.off;
   }
 }