+++ lib/src/user.dart
@@ -119,7 +119,8 @@ class User {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other is! User) return false;
-    return realmCore.userEquals(this, other);
+    //TODO: use realm_equals when https://github.com/realm/realm-core/issues/5522 is fixed
+    return id == other.id;
   }
 }