+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 * Support flexible synchronization. ([#496](https://github.com/realm/realm-dart/pull/496))
 * Added support for DateTime properties. ([#569](https://github.com/realm/realm-dart/pull/569))
 * Support session error handler. ([#577](https://github.com/realm/realm-dart/pull/577))
+* Support setting logger on AppConfiguration. ([#583](https://github.com/realm/realm-dart/pull/583))
+* Support setting logger on Realm class. Default is to print info message or worse to the console. ([#583](https://github.com/realm/realm-dart/pull/583))
 
 ### Fixed
 * Fixed an issue that would result in the wrong transaction being rolled back if you start a write transaction inside a write transaction. ([#442](https://github.com/realm/realm-dart/issues/442))